refactor: moving node specific code into separate file
